testing-requirements.txt
apipkg==1.4
attrs==18.1.0
bleach==2.1.3
bump2version==0.5.8
certifi==2018.4.16
cffi==1.11.5;
chardet==3.0.4
check-manifest==0.37
click-configfile==0.2.3
click-didyoumean==0.0.3
click-spinner==0.1.8
click==6.7
cloudsmith-api==0.26.6
cloudsmith-cli==0.5.7
cmarkgfm==0.4.2
colorama==0.3.9
configparser==3.5.0
coverage==4.5.1
docutils==0.14
execnet==1.5.0
flake8==3.5.0
future==0.16.0
html5lib==1.0.1
idna==2.6
isort==4.3.4
mccabe==0.6.1
more-itertools==4.1.0
pkginfo==1.4.2
pluggy==0.6.0
py==1.5.3
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
pygments==2.2.0
pytest-cov==2.5.1
pytest-describe==0.11.1
pytest-forked==0.2
pytest-mock==1.10.0
pytest-timeout==1.2.1
pytest-travis-fold==1.3.0
pytest-xdist==1.22.2
pytest==3.5.1
python-dateutil==2.7.2
pyyaml==3.12
readme-renderer==20.0
requests-toolbelt==0.8.0
requests==2.18.4
semver==2.7.9
simplejson==3.14.0
six==1.11.0
tqdm==4.23.3
twine==1.11.0
urllib3==1.22
webencodings==0.5.1
